requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pyodbc>=5.1.0
pandas>=2.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
